refactored types into hcl package
